[REFACTOR]: update size images
