+++ pages/_app.tsx
@@ -1,4 +1,4 @@
-import Layout from '../src/infra/components/Layout';
+import Layout from '@infra/components/Layout';
 import type { AppProps } from 'next/app';
 import '../styles/globals.css';
