+++ public/images/projects/index.ts
@@ -1,0 +1,9 @@
+const indioCoffee = '/images/projects/indio-coffee.webp';
+const notificationsServices = '/images/projects/notifications-services.webp';
+const trybeFutebolClube = '/images/projects/trybe-futebol-clube.webp';
+
+export {
+  indioCoffee,
+  notificationsServices,
+  trybeFutebolClube,
+};
