--- src/assets/projects/index.ts
@@ -1,9 +1,0 @@
-import indioCoffee from './indio-coffee.png';
-import notificationsServices from './notifications-services.png';
-import trybeFutebolClube from './trybe-futebol-clube.png';
-
-export {
-  indioCoffee,
-  notificationsServices,
-  trybeFutebolClube,
-};
