+++ tailwind.config.js
@@ -2,12 +2,12 @@
 module.exports = {
   content: [
     './pages/**/*.{js,ts,jsx,tsx}',
-    './src/infra/components/**/*.{js,ts,jsx,tsx}',
+    './src/**/*.{js,ts,jsx,tsx}',
   ],
   theme: {
     extend: {
       backgroundImage: {
-        me: 'url(\'/assets/me.jpg\')',
+        me: 'url(\'/images/me.jpg\')',
       },
     },
   },
